Feat / 마이 홈 배경 높이 짤림 현상 수정, 사용자 Faq 제목 추가 및 컴포넌트 분리, 로그인,회원가입 배경 임시 제거  #91
