+++ src/pages/Signup.styles.js
@@ -4,5 +4,6 @@ export const S = {
   Signup: styled.div`
     padding: 25px 20px;
     height: calc(100vh - 113px);
+    background: #fff;
   `,
 };
